+++ itou/templates/apply/submit/application/base.html
@@ -38,7 +38,7 @@
             <p>
                 Dernière actualisation du profil : {{ job_seeker.last_checked_at|date }} à {{ job_seeker.last_checked_at|time }}
                 {% if can_view_personal_information and not request.user.is_job_seeker %}
-                    <a class="btn btn-link" href="{% url "apply:update_job_seeker_step_1" company_pk=siae.pk job_seeker_public_id=job_seeker.public_id %}">Vérifier le profil</a>
+                    <a class="btn btn-link" href="{% url "job_seekers_views:update_job_seeker_step_1" company_pk=siae.pk job_seeker_public_id=job_seeker.public_id %}">Vérifier le profil</a>
                 {% endif %}
                 {% if new_check_needed %}<i class="ri-information-line ri-xl text-warning"></i>{% endif %}
             </p>
