+++ itou/www/job_seekers_views/views.py
@@ -2,11 +2,15 @@ import logging
 
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
+from django.core.exceptions import PermissionDenied
+from django.db import transaction
 from django.db.models import Count, DateTimeField, Exists, IntegerField, Max, OuterRef, Subquery
 from django.db.models.functions import Coalesce
 from django.forms import ValidationError
 from django.http import HttpResponseRedirect
+from django.shortcuts import get_object_or_404
 from django.urls import reverse
+from django.utils import timezone
 from django.utils.html import format_html
 from django.views.generic import DetailView, ListView
 
@@ -16,13 +20,22 @@ from itou.eligibility.models.iae import EligibilityDiagnosis
 from itou.gps.models import FollowUpGroup
 from itou.job_applications.models import JobApplication
 from itou.users.enums import UserKind
-from itou.users.models import User
+from itou.users.models import JobSeekerProfile, User
+from itou.utils.apis.exceptions import AddressLookupError
+from itou.utils.emails import redact_email_address
 from itou.utils.pagination import ItouPaginator
 from itou.utils.session import SessionNamespace, SessionNamespaceRequiredMixin
 from itou.utils.urls import get_safe_url
 from itou.www.apply.views.submit_views import ApplyStepBaseView, ApplyStepForSenderBaseView
 
-from .forms import CheckJobSeekerNirForm, FilterForm, JobSeekerExistsForm
+from .forms import (
+    CheckJobSeekerNirForm,
+    CreateOrUpdateJobSeekerStep1Form,
+    CreateOrUpdateJobSeekerStep2Form,
+    CreateOrUpdateJobSeekerStep3Form,
+    FilterForm,
+    JobSeekerExistsForm,
+)
 
 
 logger = logging.getLogger(__name__)
@@ -320,9 +333,9 @@ class SearchByEmailForSenderView(SessionNamespaceRequiredMixin, ApplyStepForSend
                 profile_infos.update({"nir": nir})
                 self.job_seeker_session.update({"user": user_infos, "profile": profile_infos})
                 view_name = (
-                    "apply:create_job_seeker_step_1_for_hire"
+                    "job_seekers_views:create_job_seeker_step_1_for_hire"
                     if self.hire_process
-                    else "apply:create_job_seeker_step_1_for_sender"
+                    else "job_seekers_views:create_job_seeker_step_1_for_sender"
                 )
 
                 return HttpResponseRedirect(
@@ -387,3 +400,592 @@ class SearchByEmailForSenderView(SessionNamespaceRequiredMixin, ApplyStepForSend
             "siae": self.company,
             "preview_mode": False,
         }
+
+
+class CreateJobSeekerForSenderBaseView(SessionNamespaceRequiredMixin, ApplyStepForSenderBaseView):
+    required_session_namespaces = ["job_seeker_session"]
+
+    def __init__(self):
+        super().__init__()
+        self.job_seeker_session = None
+
+    def setup(self, request, *args, **kwargs):
+        self.job_seeker_session = SessionNamespace(request.session, kwargs["session_uuid"])
+        super().setup(request, *args, **kwargs)
+
+    def get_back_url(self):
+        view_name = self.previous_hire_url if self.hire_process else self.previous_apply_url
+        return reverse(
+            view_name,
+            kwargs={"company_pk": self.company.pk, "session_uuid": self.job_seeker_session.name},
+        ) + ("?gps=true" if self.is_gps else "")
+
+    def get_next_url(self):
+        view_name = self.next_hire_url if self.hire_process else self.next_apply_url
+        return reverse(
+            view_name,
+            kwargs={"company_pk": self.company.pk, "session_uuid": self.job_seeker_session.name},
+        ) + ("?gps=true" if self.is_gps else "")
+
+
+class CreateJobSeekerStep1ForSenderView(CreateJobSeekerForSenderBaseView):
+    template_name = "job_seekers_views/create_or_update_job_seeker/step_1.html"
+
+    previous_apply_url = "job_seekers_views:search_by_email_for_sender"
+    previous_hire_url = "job_seekers_views:search_by_email_for_hire"
+    next_apply_url = "job_seekers_views:create_job_seeker_step_2_for_sender"
+    next_hire_url = "job_seekers_views:create_job_seeker_step_2_for_hire"
+
+    def __init__(self):
+        super().__init__()
+        self.form = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        session_birthdate = self.job_seeker_session.get("profile", {}).get(
+            "birthdate", self.job_seeker_session.get("user", {}).get("birthdate")
+        )  # TODO(xfernandez): drop fallback on self.job_seeker_session["user"] in a week
+        session_nir = self.job_seeker_session.get("profile", {}).get("nir")
+        session_lack_of_nir_reason = self.job_seeker_session.get("profile", {}).get("lack_of_nir_reason")
+        session_birth_place = self.job_seeker_session.get("profile", {}).get("birth_place")
+        session_birth_country = self.job_seeker_session.get("profile", {}).get("birth_country")
+
+        self.form = CreateOrUpdateJobSeekerStep1Form(
+            data=request.POST or None,
+            initial=self.job_seeker_session.get("user", {})
+            | {
+                "birthdate": session_birthdate,
+                "nir": session_nir,
+                "lack_of_nir_reason": session_lack_of_nir_reason,
+                "birth_place": session_birth_place,
+                "birth_country": session_birth_country,
+            },
+        )
+
+    def post(self, request, *args, **kwargs):
+        context = self.get_context_data(**kwargs)
+        context["confirmation_needed"] = False
+        if self.form.is_valid():
+            existing_job_seeker = User.objects.filter(
+                kind=UserKind.JOB_SEEKER,
+                jobseeker_profile__birthdate=self.form.cleaned_data["birthdate"],
+                first_name__unaccent__iexact=self.form.cleaned_data["first_name"],
+                last_name__unaccent__iexact=self.form.cleaned_data["last_name"],
+            ).first()
+            if existing_job_seeker and not self.form.data.get("confirm"):
+                # If an existing job seeker matches the info, a confirmation is required
+                context["confirmation_needed"] = True
+                context["redacted_existing_email"] = redact_email_address(existing_job_seeker.email)
+                context["email_to_create"] = self.job_seeker_session.get("user", {}).get("email", "")
+
+            if not context["confirmation_needed"]:
+                self.job_seeker_session.set(
+                    "user",
+                    self.job_seeker_session.get("user", {}) | self.form.cleaned_data_without_profile_fields,
+                )
+                self.job_seeker_session.set(
+                    "profile",
+                    self.job_seeker_session.get("profile", {}) | self.form.cleaned_data_from_profile_fields,
+                )
+                return HttpResponseRedirect(self.get_next_url())
+
+        return self.render_to_response(context)
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "form": self.form,
+            "matomo_form_name": "apply-create-job-seeker-identity",
+            "progress": "20",
+        }
+
+
+class CreateJobSeekerStep2ForSenderView(CreateJobSeekerForSenderBaseView):
+    template_name = "job_seekers_views/create_or_update_job_seeker/step_2.html"
+
+    previous_apply_url = "job_seekers_views:create_job_seeker_step_1_for_sender"
+    previous_hire_url = "job_seekers_views:create_job_seeker_step_1_for_hire"
+    next_apply_url = "job_seekers_views:create_job_seeker_step_3_for_sender"
+    next_hire_url = "job_seekers_views:create_job_seeker_step_3_for_hire"
+
+    def __init__(self):
+        super().__init__()
+        self.form = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        self.form = CreateOrUpdateJobSeekerStep2Form(
+            data=request.POST or None, initial=self.job_seeker_session.get("user", {})
+        )
+
+    def post(self, request, *args, **kwargs):
+        if self.form.is_valid():
+            self.job_seeker_session.set("user", self.job_seeker_session.get("user") | self.form.cleaned_data)
+            return HttpResponseRedirect(self.get_next_url())
+
+        return self.render_to_response(self.get_context_data(**kwargs))
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "form": self.form,
+            "progress": "40",
+        }
+
+
+class CreateJobSeekerStep3ForSenderView(CreateJobSeekerForSenderBaseView):
+    template_name = "job_seekers_views/create_or_update_job_seeker/step_3.html"
+
+    previous_apply_url = "job_seekers_views:create_job_seeker_step_2_for_sender"
+    previous_hire_url = "job_seekers_views:create_job_seeker_step_2_for_hire"
+    next_apply_url = "job_seekers_views:create_job_seeker_step_end_for_sender"
+    next_hire_url = "job_seekers_views:create_job_seeker_step_end_for_hire"
+
+    def __init__(self):
+        super().__init__()
+        self.form = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+
+        self.form = CreateOrUpdateJobSeekerStep3Form(
+            data=request.POST or None,
+            initial=self.job_seeker_session.get("profile", {}),
+        )
+
+    def post(self, request, *args, **kwargs):
+        if self.form.is_valid():
+            self.job_seeker_session.set("profile", self.job_seeker_session.get("profile", {}) | self.form.cleaned_data)
+            return HttpResponseRedirect(self.get_next_url())
+
+        return self.render_to_response(self.get_context_data(**kwargs))
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "form": self.form,
+            "progress": "60",
+        }
+
+
+class CreateJobSeekerStepEndForSenderView(CreateJobSeekerForSenderBaseView):
+    template_name = "job_seekers_views/create_or_update_job_seeker/step_end.html"
+
+    previous_apply_url = "job_seekers_views:create_job_seeker_step_3_for_sender"
+    previous_hire_url = "job_seekers_views:create_job_seeker_step_3_for_hire"
+    next_apply_url = "apply:application_jobs"
+    next_hire_url = None  # Depends on GEIQ/non-GEIQ
+
+    def __init__(self):
+        super().__init__()
+        self.profile = None
+
+    def _get_user_data_from_session(self):
+        return {
+            k: v
+            for k, v in self.job_seeker_session.get("user").items()
+            if k
+            not in [
+                "lack_of_nir",
+                # Address autocomplete fields
+                "fill_mode",
+                "address_for_autocomplete",
+                "insee_code",
+                # TODO(xfernandez): remove birthdate in a week
+                "birthdate",
+            ]
+        }
+
+    def _get_profile_data_from_session(self):
+        fields_to_exclude = [
+            # Dummy fields used by CreateOrUpdateJobSeekerStep3Form()
+            "pole_emploi",
+            "pole_emploi_id_forgotten",
+            "rsa_allocation",
+            "unemployed",
+            "ass_allocation",
+            "aah_allocation",
+            "lack_of_nir",
+            # ForeignKeys - the session value will be the ID serialization and not the instance
+            "birth_place",
+            "birth_country",
+        ]
+
+        birth_data = {
+            "birth_place_id": self.job_seeker_session.get("profile", {}).get("birth_place"),
+            "birth_country_id": self.job_seeker_session.get("profile", {}).get("birth_country"),
+        }
+
+        # TODO(xfernandez): remove user session birthdate handling in a week
+        if birthdate_from_user_session := self.job_seeker_session.get("user", {}).get("birthdate"):
+            birth_data |= {"birthdate": birthdate_from_user_session}
+
+        return birth_data | {
+            k: v for k, v in self.job_seeker_session.get("profile").items() if k not in fields_to_exclude
+        }
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+
+        self.profile = JobSeekerProfile(
+            user=User(**self._get_user_data_from_session()),
+            **self._get_profile_data_from_session(),
+        )
+
+    def get_next_url(self):
+        kwargs = {"company_pk": self.company.pk, "job_seeker_public_id": self.profile.user.public_id}
+
+        if self.is_gps:
+            kwargs = {}
+            view_name = "gps:my_groups"
+        elif self.hire_process:
+            if self.company.kind == CompanyKind.GEIQ:
+                view_name = "apply:geiq_eligibility_for_hire"
+            else:
+                view_name = "apply:eligibility_for_hire"
+        else:
+            view_name = self.next_apply_url
+
+        return reverse(
+            view_name,
+            kwargs=kwargs,
+        )
+
+    def post(self, request, *args, **kwargs):
+        try:
+            with transaction.atomic():
+                user = User.create_job_seeker_by_proxy(
+                    self.sender, **self._get_user_data_from_session(), acting_organization=request.current_organization
+                )
+                self.profile = user.jobseeker_profile
+                for k, v in self._get_profile_data_from_session().items():
+                    setattr(self.profile, k, v)
+                self.profile.save()
+        except ValidationError as e:
+            messages.error(request, " ".join(e.messages))
+            url = reverse("dashboard:index")
+        else:
+            try:
+                user.geocode_address()
+            except AddressLookupError:
+                # Nothing to do: re-raised and already logged as error
+                pass
+            else:
+                user.save()
+
+            self.job_seeker_session.delete()
+            url = self.get_next_url()
+
+            if self.is_gps:
+                FollowUpGroup.objects.follow_beneficiary(beneficiary=user, user=request.user, is_referent=True)
+
+        return HttpResponseRedirect(url)
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {"profile": self.profile, "progress": "80"}
+
+
+class UpdateJobSeekerBaseView(SessionNamespaceRequiredMixin, ApplyStepBaseView):
+    def __init__(self):
+        super().__init__()
+        self.job_seeker_session = None
+
+    def get_job_seeker_queryset(self):
+        return User.objects.filter(kind=UserKind.JOB_SEEKER)
+
+    def setup(self, request, *args, **kwargs):
+        self.job_seeker = get_object_or_404(self.get_job_seeker_queryset(), public_id=kwargs["job_seeker_public_id"])
+        self.job_seeker_session = SessionNamespace(request.session, f"job_seeker-{self.job_seeker.public_id}")
+        if request.user.is_authenticated and (
+            request.user.is_job_seeker or not request.user.can_view_personal_information(self.job_seeker)
+        ):
+            # Since the link leading to this process isn't visible to those users, this should never happen
+            raise PermissionDenied("Votre utilisateur n'est pas autorisé à vérifier les informations de ce candidat")
+        super().setup(request, *args, **kwargs)
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "update_job_seeker": True,
+            "job_seeker": self.job_seeker,
+            "step_3_url": reverse(
+                "job_seekers_views:update_job_seeker_step_3_for_hire"
+                if self.hire_process
+                else "job_seekers_views:update_job_seeker_step_3",
+                kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+            ),
+            "reset_url": reverse(
+                "apply:application_jobs",
+                kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+            ),
+            "readonly_form": False,
+        }
+
+    def _disable_form(self):
+        for field in self.form:
+            field.field.disabled = True
+
+    def get_back_url(self):
+        view_name = self.previous_hire_url if self.hire_process else self.previous_apply_url
+        return reverse(
+            view_name,
+            kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+        )
+
+    def get_next_url(self):
+        view_name = self.next_hire_url if self.hire_process else self.next_apply_url
+        return reverse(
+            view_name,
+            kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+        )
+
+
+class UpdateJobSeekerStep1View(UpdateJobSeekerBaseView):
+    template_name = "job_seekers_views/create_or_update_job_seeker/step_1.html"
+
+    previous_apply_url = "apply:application_jobs"
+    previous_hire_url = "apply:check_job_seeker_info_for_hire"
+    next_apply_url = "job_seekers_views:update_job_seeker_step_2"
+    next_hire_url = "job_seekers_views:update_job_seeker_step_2_for_hire"
+
+    def __init__(self):
+        super().__init__()
+        self.form = None
+
+    def get_job_seeker_queryset(self):
+        return super().get_job_seeker_queryset().select_related("jobseeker_profile")
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        if not request.user.is_authenticated:
+            # Do nothing, LoginRequiredMixin will raise in dispatch()
+            return
+        if not self.job_seeker_session.exists():
+            self.job_seeker_session.init({"user": {}})
+        session_nir = self.job_seeker_session.get("profile", {}).get("nir")
+        session_lack_of_nir_reason = self.job_seeker_session.get("profile", {}).get("lack_of_nir_reason")
+
+        self.form = CreateOrUpdateJobSeekerStep1Form(
+            instance=self.job_seeker,
+            initial=self.job_seeker_session.get("user", {})
+            | {
+                "nir": session_nir if session_nir is not None else self.job_seeker.jobseeker_profile.nir,
+                "lack_of_nir_reason": (
+                    session_lack_of_nir_reason
+                    if session_lack_of_nir_reason is not None
+                    else self.job_seeker.jobseeker_profile.lack_of_nir_reason
+                ),
+            },
+            data=request.POST or None,
+        )
+        if not self.request.user.can_edit_personal_information(self.job_seeker):
+            self._disable_form()
+
+    def post(self, request, *args, **kwargs):
+        if not self.request.user.can_edit_personal_information(self.job_seeker):
+            return HttpResponseRedirect(self.get_next_url())
+        if self.form.is_valid():
+            self.job_seeker_session.set(
+                "user",
+                self.job_seeker_session.get("user", {}) | self.form.cleaned_data_without_profile_fields,
+            )
+            self.job_seeker_session.set(
+                "profile",
+                self.job_seeker_session.get("profile", {}) | self.form.cleaned_data_from_profile_fields,
+            )
+            return HttpResponseRedirect(self.get_next_url())
+
+        return self.render_to_response(self.get_context_data(**kwargs))
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "confirmation_needed": False,
+            "form": self.form,
+            "matomo_form_name": "apply-update-job-seeker-identity",
+            "readonly_form": not self.request.user.can_edit_personal_information(self.job_seeker),
+            "progress": "20",
+        }
+
+
+class UpdateJobSeekerStep2View(UpdateJobSeekerBaseView):
+    template_name = "job_seekers_views/create_or_update_job_seeker/step_2.html"
+    required_session_namespaces = ["job_seeker_session"] + UpdateJobSeekerBaseView.required_session_namespaces
+
+    previous_apply_url = "job_seekers_views:update_job_seeker_step_1"
+    previous_hire_url = "job_seekers_views:update_job_seeker_step_1_for_hire"
+    next_apply_url = "job_seekers_views:update_job_seeker_step_3"
+    next_hire_url = "job_seekers_views:update_job_seeker_step_3_for_hire"
+
+    def __init__(self):
+        super().__init__()
+        self.form = None
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+        self.form = CreateOrUpdateJobSeekerStep2Form(
+            instance=self.job_seeker,
+            initial=self.job_seeker_session.get("user", {}),
+            data=request.POST or None,
+        )
+        if not self.request.user.can_edit_personal_information(self.job_seeker):
+            self._disable_form()
+
+    def post(self, request, *args, **kwargs):
+        if not self.request.user.can_edit_personal_information(self.job_seeker):
+            return HttpResponseRedirect(self.get_next_url())
+        if self.form.is_valid():
+            self.job_seeker_session.set("user", self.job_seeker_session.get("user") | self.form.cleaned_data)
+            return HttpResponseRedirect(self.get_next_url())
+
+        return self.render_to_response(self.get_context_data(**kwargs))
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "form": self.form,
+            "readonly_form": not self.request.user.can_edit_personal_information(self.job_seeker),
+            "progress": "40",
+        }
+
+
+class UpdateJobSeekerStep3View(UpdateJobSeekerBaseView):
+    template_name = "job_seekers_views/create_or_update_job_seeker/step_3.html"
+    required_session_namespaces = ["job_seeker_session"] + UpdateJobSeekerBaseView.required_session_namespaces
+
+    previous_apply_url = "job_seekers_views:update_job_seeker_step_2"
+    previous_hire_url = "job_seekers_views:update_job_seeker_step_2_for_hire"
+    next_apply_url = "job_seekers_views:update_job_seeker_step_end"
+    next_hire_url = "job_seekers_views:update_job_seeker_step_end_for_hire"
+
+    def __init__(self):
+        super().__init__()
+        self.form = None
+
+    def get_job_seeker_queryset(self):
+        return super().get_job_seeker_queryset().select_related("jobseeker_profile")
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+
+        session_pole_emploi_id = self.job_seeker_session.get("profile", {}).get("pole_emploi_id")
+        session_lack_of_pole_emploi_id_reason = self.job_seeker_session.get("profile", {}).get(
+            "lack_of_pole_emploi_id_reason"
+        )
+        initial_form_data = self.job_seeker_session.get("profile", {}) | {
+            "pole_emploi_id": (
+                session_pole_emploi_id
+                if session_pole_emploi_id is not None
+                else self.job_seeker.jobseeker_profile.pole_emploi_id
+            ),
+            "lack_of_pole_emploi_id_reason": (
+                session_lack_of_pole_emploi_id_reason
+                if session_lack_of_pole_emploi_id_reason is not None
+                else self.job_seeker.jobseeker_profile.lack_of_pole_emploi_id_reason
+            ),
+        }
+        self.form = CreateOrUpdateJobSeekerStep3Form(
+            instance=self.job_seeker.jobseeker_profile if self.job_seeker.has_jobseeker_profile else None,
+            initial=initial_form_data,
+            data=request.POST or None,
+        )
+
+    def post(self, request, *args, **kwargs):
+        if self.form.is_valid():
+            self.job_seeker_session.set("profile", self.job_seeker_session.get("profile", {}) | self.form.cleaned_data)
+            return HttpResponseRedirect(self.get_next_url())
+
+        return self.render_to_response(self.get_context_data(**kwargs))
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {
+            "form": self.form,
+            "progress": "60",
+        }
+
+
+class UpdateJobSeekerStepEndView(UpdateJobSeekerBaseView):
+    template_name = "job_seekers_views/create_or_update_job_seeker/step_end.html"
+    required_session_namespaces = ["job_seeker_session"] + UpdateJobSeekerBaseView.required_session_namespaces
+
+    previous_apply_url = "job_seekers_views:update_job_seeker_step_3"
+    previous_hire_url = "job_seekers_views:update_job_seeker_step_3_for_hire"
+    next_apply_url = "apply:application_jobs"
+    next_hire_url = "apply:check_job_seeker_info_for_hire"
+
+    def __init__(self):
+        super().__init__()
+        self.profile = None
+        self.updated_user_fields = []
+
+    def _get_profile_data_from_session(self):
+        fields_to_exclude = [
+            # Dummy fields used by CreateOrUpdateJobSeekerStep3Form()
+            "pole_emploi",
+            "pole_emploi_id_forgotten",
+            "rsa_allocation",
+            "unemployed",
+            "ass_allocation",
+            "aah_allocation",
+            "lack_of_nir",
+            # ForeignKeys - the session value will be the ID serialization and not the instance
+            "birth_place",
+            "birth_country",
+        ]
+
+        birth_data = {
+            "birth_place_id": self.job_seeker_session.get("profile", {}).get("birth_place"),
+            "birth_country_id": self.job_seeker_session.get("profile", {}).get("birth_country"),
+        }
+
+        return birth_data | {
+            k: v for k, v in self.job_seeker_session.get("profile", {}).items() if k not in fields_to_exclude
+        }
+
+    def setup(self, request, *args, **kwargs):
+        super().setup(request, *args, **kwargs)
+
+        allowed_user_fields_to_update = []
+        if self.request.user.can_edit_personal_information(self.job_seeker):
+            allowed_user_fields_to_update.extend(CreateOrUpdateJobSeekerStep1Form.Meta.fields)
+            allowed_user_fields_to_update.extend(CreateOrUpdateJobSeekerStep2Form.Meta.fields)
+
+        for field in allowed_user_fields_to_update:
+            if field in self.job_seeker_session.get("user", {}):
+                session_value = self.job_seeker_session.get("user")[field]
+                if session_value != getattr(self.job_seeker, field):
+                    setattr(self.job_seeker, field, session_value)
+                    self.updated_user_fields.append(field)
+
+        if not self.job_seeker.has_jobseeker_profile:
+            self.profile = JobSeekerProfile(
+                user=self.job_seeker,
+                **self._get_profile_data_from_session(),
+            )
+        else:
+            self.profile = self.job_seeker.jobseeker_profile
+            for k, v in self._get_profile_data_from_session().items():
+                setattr(self.profile, k, v)
+
+    @transaction.atomic
+    def post(self, request, *args, **kwargs):
+        self.updated_user_fields.append("last_checked_at")
+        try:
+            if "address_line_1" in self.updated_user_fields or "post_code" in self.updated_user_fields:
+                try:
+                    self.job_seeker.geocode_address()
+                except AddressLookupError:
+                    # Nothing to do: re-raised and already logged as error
+                    pass
+                else:
+                    self.updated_user_fields.extend(["coords", "geocoding_score"])
+            self.job_seeker.last_checked_at = timezone.now()
+            self.job_seeker.save(update_fields=self.updated_user_fields)
+        except ValidationError as e:
+            messages.error(request, " ".join(e.messages))
+            url = reverse(
+                "job_seekers_views:update_job_seeker_step_1",
+                kwargs={"company_pk": self.company.pk, "job_seeker_public_id": self.job_seeker.public_id},
+            )
+        else:
+            self.profile.save()
+            self.job_seeker_session.delete()
+            url = self.get_next_url()
+        return HttpResponseRedirect(url)
+
+    def get_context_data(self, **kwargs):
+        return super().get_context_data(**kwargs) | {"profile": self.profile, "progress": "80"}
