+++ tests/gps/test_create_beneficiary.py
@@ -92,7 +92,7 @@ def test_create_job_seeker(_mock, client):
     }
     next_url = (
         reverse(
-            "apply:create_job_seeker_step_1_for_sender",
+            "job_seekers_views:create_job_seeker_step_1_for_sender",
             kwargs={"company_pk": singleton.pk, "session_uuid": job_seeker_session_name},
         )
         + "?gps=true"
@@ -139,7 +139,7 @@ def test_create_job_seeker(_mock, client):
 
     next_url = (
         reverse(
-            "apply:create_job_seeker_step_2_for_sender",
+            "job_seekers_views:create_job_seeker_step_2_for_sender",
             kwargs={"company_pk": singleton.pk, "session_uuid": job_seeker_session_name},
         )
         + "?gps=true"
@@ -163,7 +163,7 @@ def test_create_job_seeker(_mock, client):
 
     next_url = (
         reverse(
-            "apply:create_job_seeker_step_3_for_sender",
+            "job_seekers_views:create_job_seeker_step_3_for_sender",
             kwargs={"company_pk": singleton.pk, "session_uuid": job_seeker_session_name},
         )
         + "?gps=true"
@@ -197,7 +197,7 @@ def test_create_job_seeker(_mock, client):
 
     next_url = (
         reverse(
-            "apply:create_job_seeker_step_end_for_sender",
+            "job_seekers_views:create_job_seeker_step_end_for_sender",
             kwargs={"company_pk": singleton.pk, "session_uuid": job_seeker_session_name},
         )
         + "?gps=true"
